+++ NEWS.md
@@ -3,5 +3,6 @@
 * Create [RAML file](ramls/ldp.raml), and associated [JSON Schemas and example files](ramls), describing the WSAPI. Fixes #5.
 * Write [WSAPI overview documentation](ramls/overview.md). Fixes #6.
 * Set up automatic generation of WSAPI documentation. Fixes #8.
+* Support `limit` parameter in query. Fixes #7.
 * Initial release